membuat informasi totalData sama  dengan total isDeleted == false

+++ src/main/java/com/latihan/latihanmobil/DTO/ResponseMobilDTO.java
@@ -12,5 +12,5 @@ public class ResponseMobilDTO {
     private List<MobilDTO> mobilDTO;
     private Boolean loadSuccess;
     private Integer totalPage;
-    private Long totalData;
+    private Integer totalData;
 }
